arm64: dts: qcom: Add device tree for Samsung Galaxy Tab A (2015)
This dts adds support for wifi and lte variants of
Samsung Galaxy Tab A tablets released in 2015. (with both 8.0 and
9.7 inch displays)

Add a device tree for gt58 and gt510 with initial support for:

- SDHCI (internal and external storage)
- USB Device Mode
- UART
- Regulators
- WCNSS (WiFi/BT)
- GPIO keys

Most of the device tree is separated into
msm8916-samsung-gt5-common.dtsi as these devices share a lot of
hardware with each other.

Signed-off-by: Jasper Korten <jja2000@gmail.com>
Co-developed-by: Siddharth Manthan <siddharth_manthan@outlook.com>
Signed-off-by: Siddharth Manthan <siddharth_manthan@outlook.com>
Co-developed-by: Nikita Travkin <nikita@trvn.ru>
Signed-off-by: Nikita Travkin <nikita@trvn.ru>